WIP: boards: amlogic: add WeTek Play2 support
WeTek Play2 is a P200 based design with S905-H chip. Board config is the
same as P200 except for device specific strings. WeTek has not published
full u-boot sources, but has shared them so that the FIP binaries can be
extracted and hosted/shared for use in the FIP recipe.